Fixed Multiplayer integration
It seems that the Multiplayer integration has broken due to Traverse no longer being able to find the method. This happens due to an ambiguous match, as 2 methods share the same name. Considering that it seems Multiplayer added both of the methods at the same time, it's likely that Traverse stopped handling ambiguous matches like those to prevent potential issues.

The fix for the issue was fairly simple by providing the method argument types, so the correct method could be retrieved by Traverse. After some testing everything seems to work.

On top of fixing the issue, I've added additional check to see if the method was found by calling `.MethodExists()` (which is just a null check done by Traverse).